# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Demo API
  version: 0.1.0
servers:
  - url: http://localhost:8080
paths:
  /api/health:
    get:
      tags:
        - health
      operationId: getHealth
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
  /api/quotes:
    post:
      tags:
        - quotes
      operationId: createQuote
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuoteRequest"
      responses:
        "200":
          description: Quote created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuoteResponse"
        "400":
          description: Driver ineligible (age or postcode)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/policies:
    post:
      tags:
        - policies
      operationId: bindPolicy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BindRequest"
      responses:
        "201":
          description: Policy bound
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Policy"
        "404":
          description: Quote not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/policies/{id}:
    get:
      tags:
        - policies
      operationId: getPolicy
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Policy retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Policy"
        "404":
          description: Policy not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthResponse:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          example: ok
    QuoteRequest:
      type: object
      required:
        - driverName
        - dateOfBirth
        - licenceNumber
        - postcode
        - vehicleRef
      properties:
        driverName:
          type: string
          example: Alex Driver
        dateOfBirth:
          type: string
          format: date
          example: "1995-06-12"
        licenceNumber:
          type: string
          example: DRIVE701054AB9XY
        postcode:
          type: string
          example: E14 5AB
        vehicleRef:
          type: string
          description: Vehicle registration plate or VIN
          example: AB12 CDE
    QuoteResponse:
      type: object
      required:
        - quoteId
        - premium
      properties:
        quoteId:
          type: string
          example: 65f1a2c3d4e5f60718293a4b
        premium:
          type: number
          format: double
          example: 550.00
    BindRequest:
      type: object
      required:
        - quoteId
      properties:
        quoteId:
          type: string
          example: 65f1a2c3d4e5f60718293a4b
    Policy:
      type: object
      required:
        - policyId
        - quoteId
        - premium
        - effectiveFrom
        - effectiveTo
      properties:
        policyId:
          type: string
          example: 65f1aabbccddeeff00112233
        quoteId:
          type: string
          example: 65f1a2c3d4e5f60718293a4b
        premium:
          type: number
          format: double
          example: 550.00
        effectiveFrom:
          type: string
          format: date
          example: "2026-05-17"
        effectiveTo:
          type: string
          format: date
          example: "2027-05-17"
    ErrorResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          example: Postcode not supported